% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sef.R
\name{read_sef}
\alias{read_sef}
\title{Read data files in Station Exchange Format version 0.2.1}
\usage{
read_sef(file = file.choose())
}
\arguments{
\item{file}{Character string giving the path of the SEF file.}
}
\value{
A data frame with seven or eigth variables (depending on time
resolution): variable code, year, month, day, (time), value,
time flag, metadata.
}
\description{
Read data files in Station Exchange Format version 0.2.1
}
\author{
Yuri Brugnara
}
