% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Variables}
\alias{Variables}
\title{List of standard variable codes}
\format{A data frame with two variables}
\source{
C3S Data Rescue Service
}
\usage{
Variables
}
\description{
List of standard variable codes
}
\keyword{datasets}
