% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sef.R
\name{write_sef}
\alias{write_sef}
\title{Write data in Station Exchange Format version 0.2.1}
\usage{
write_sef(Data, outpath, cod, nam, lat = NA, lon = NA, alt = NA,
  sou = NA, repo = NA, units = NA, metaHead = "", meta = "",
  timef = NA, note = "")
}
\arguments{
\item{Data}{A data frame with five or six variables (depending on time
resolution): variable code, year, month, day, (time in HHMM or HH:MM), value.}

\item{outpath}{Character string giving the output path (note that the filename
is generated from the source identifier, station code, start and end dates,
and variable code).}

\item{cod}{Station code.}

\item{nam}{Station name.}

\item{lat}{Station latitude (degrees North in decimal).}

\item{lon}{Station longitude (degreees East in decimal).}

\item{alt}{Station altitude (metres).}

\item{sou}{Character string giving the source identifier.}

\item{repo}{Character string giving the repository identifier.}

\item{units}{Character string giving the units.}

\item{metaHead}{Character string giving any metadata for the header.}

\item{meta}{Character vector with length equal to the number of rows
of Data, giving any metadata for the single observations.}

\item{timef}{Integer or integer vector giving the observation time period code.
If NA (the default), it will be guessed from the dimension of Data and
the variable code.}

\item{note}{Character string to be added to the end of the filename.
It will be separated from the rest of the name by an underscore.
Blanks will be also replaced by underscores.}
}
\description{
Write data in Station Exchange Format version 0.2.1
}
\author{
Yuri Brugnara
}
